A working, simple 4 track model before I screw anything up
I tried our 3D idea and it is terrible: super messy and not playable at
all. Also the ends of the four tracks intersected in a super ugly way..
just hot mess in general. So I swtiched to boring 2D and horizontal
layout. It's stereotyped but safe..decide to save this first before I
try anything else that will screw it up.